CI: Use same tag of dependency images
This ensures that if we're building from a non-main branch, we use the
appropriate tag of the dependency images during the build of downstream
images. E.g. when building the arago-build container for a 'next'
branch, it should be based on the 'next' tag of the poky-build container
instead of the (default) 'latest' tag.

Signed-off-by: Paul Barker <paul.barker@sancloud.com>