CI: Fix pre-commit caching & use dev23 image
Since we now allow concurrent CI builds, CI_PROJECT_DIR can be either
`/builds/.../0/...` or `/builds/.../1/...`. This causes an issue when
the pre-commit cache is saved by a build using one of those paths and
then restored for a build using the other path. The pre-commit cache
stores absolute paths to the cached repositories instead of relative
paths and so they refer to the alternative CI_PROJECT_DIR path, leading
to failure as the files are no longer accessible in that path.

The exact failure from a recent build is:

    $ pre-commit run -a
    An error has occurred: InvalidManifestError:
    =====> /builds/sF4Cx6V6/0/bsp/build-containers/.cache/pre-commit/repoo0jj6epp/.pre-commit-hooks.yaml is not a file

To prevent these errors, we leave the pre-commit cache directory in the
default path of `~/.cache/pre-commit` (which is always the same absolute
path) and symlink this to a cache folder inside the project directory so
that we can cache it.

Note that the cached directory must live inside CI_PROJECT_DIR unless
https://gitlab.com/gitlab-org/gitlab/-/issues/14151 is fixed upstream.

Signed-off-by: Paul Barker <paul.barker@sancloud.com>